requirements.txt
fastapi
uvicorn
numpy
tensorflow
pillow
python-multipart
scikit-learn